Omstruktuerering av fyllut backend. Trekke ut logikk for hvert endepunkt, og plasser de under routers-mappen. Eksponere en del av endepunktene også under /api slik at vi gradvis kan gå over til disse; dette for å tydeliggjøre hva som er tjenester, og hva som returnerer statiske ressurseri, og minske sjansen for at det håndteres feilaktig.
