+++ packages/fyllut/server/routers/api/countries.js
@@ -1,0 +1,7 @@
+import { getCountries } from "../../utils/countries.js";
+
+const countries = {
+  get: (req, res) => res.json(getCountries(req.query.lang)),
+};
+
+export default countries;
