+++ packages/fyllut/server/routers/api/mottaksadresser.js
@@ -1,0 +1,16 @@
+import { config } from "../../config/config.js";
+import { fetchFromFormioApi, loadFileFromDirectory } from "../../utils/forms.js";
+
+const { useFormioApi, formioProjectUrl, resourcesDir } = config;
+
+const loadMottaksadresser = async () => {
+  return useFormioApi
+    ? await fetchFromFormioApi(`${formioProjectUrl}/mottaksadresse/submission`)
+    : await loadFileFromDirectory(resourcesDir, "mottaksadresser.json", []);
+};
+
+const mottaksadresser = {
+  get: async (req, res) => res.json(await loadMottaksadresser()),
+};
+
+export default mottaksadresser;
